Migrate most of summary related code from collection_services and exp_services to summary_services
